Refactored stopwatch
